+++ packages/conventional-changelog-cli/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.12](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog-cli@2.0.11...conventional-changelog-cli@2.0.12) (2019-02-14)
+
+**Note:** Version bump only for package conventional-changelog-cli
+
+
+
+
+
 ## [2.0.11](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog-cli@2.0.10...conventional-changelog-cli@2.0.11) (2018-11-01)
 
 **Note:** Version bump only for package conventional-changelog-cli