+++ packages/conventional-changelog/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.6](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog@3.0.5...conventional-changelog@3.0.6) (2019-02-14)
+
+**Note:** Version bump only for package conventional-changelog
+
+
+
+
+
 ## [3.0.5](https://github.com/conventional-changelog/conventional-changelog/compare/conventional-changelog@3.0.4...conventional-changelog@3.0.5) (2018-11-01)
 
 **Note:** Version bump only for package conventional-changelog